chore: prepare VIP-660 for proposal
Renumber vip-665 -> vip-660 (on-chain proposalCount 659), rework the
description around the verified on-chain funding picture, add References.
